simplify encryption stream
